scripts/requirements.txt
numpy==2.2.5
pandas==2.2.3
PyYAML==6.0.2
PyYAML==6.0.2
Requests==2.32.3
